webgl: polling loop fixes, and put the tests where they belong
EVENT_PUMP_INTERVAL becomes EVENT_POLL_INTERVAL: the loop dispatches nothing
itself, it makes a no-op round-trip so Playwright dispatches what it queued.

A polling failure no longer closes the viewer. The except fell through to
_cleanup(), so one transient failure destroyed the browser mid-session; it now
parks until shutdown and warns.

filter_webgl_failures' unit test and the delivery test move to
test_headless.py, whose subject they are.
